Remove '.internal' folder, moving contents up
ESLint ignores everything in folders that start with '.' (god knows why this was decided). Although it is possible to force ESLint via the command line to not ignore these folders, its currently not possible to easily setup a way for the ESLint plugin in VS Code to do the same, so dropping the '.internal' folder is much easier, however disapointing it is.